refactor: address PR review comments
See https://github.com/ibi-group/otp-middleware/pull/119
