+++ src/test/java/org/opentripplanner/middleware/OtpMiddlewareTest.java
@@ -16,7 +16,7 @@ import java.net.http.HttpResponse;
 import java.nio.charset.StandardCharsets;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.opentripplanner.middleware.testUtils.CommonTestUtils.isEndToEnd;
+import static org.opentripplanner.middleware.testutils.CommonTestUtils.IS_END_TO_END;
 import static org.opentripplanner.middleware.utils.ConfigUtils.isRunningCi;
 
 /**
@@ -45,7 +45,7 @@ public class OtpMiddlewareTest {
         // If in the e2e environment, use the secret env.yml file to start the server.
         // TODO: When ran on Travis CI, this file will automatically be setup.
         String[] args;
-        if (isEndToEnd) {
+        if (IS_END_TO_END) {
             // Check if running in Travis CI. If so, use Travis environment variables instead of config file.
             args = isRunningCi ? new String[]{} : new String[]{"configurations/default/env.yml"};
         } else {
