+++ src/test/java/org/opentripplanner/middleware/auth/Auth0ConnectionTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.middleware.controllers.api;
+package org.opentripplanner.middleware.auth;
 
 import com.auth0.exception.Auth0Exception;
 import com.auth0.json.mgmt.users.User;
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.opentripplanner.middleware.OtpMiddlewareTest;
-import org.opentripplanner.middleware.auth.Auth0Connection;
-import org.opentripplanner.middleware.auth.Auth0Users;
 import org.opentripplanner.middleware.models.AbstractUser;
 import org.opentripplanner.middleware.models.ApiUser;
 import org.opentripplanner.middleware.models.OtpUser;
@@ -27,10 +25,10 @@ import static org.junit.jupiter.api.Assumptions.assumeTrue;
 import static org.opentripplanner.middleware.auth.Auth0Connection.*;
 import static org.opentripplanner.middleware.controllers.api.ApiUserController.API_USER_PATH;
 import static org.opentripplanner.middleware.controllers.api.OtpUserController.OTP_USER_PATH;
-import static org.opentripplanner.middleware.testUtils.ApiTestUtils.TEMP_AUTH0_USER_PASSWORD;
-import static org.opentripplanner.middleware.testUtils.ApiTestUtils.mockAuthenticatedGet;
-import static org.opentripplanner.middleware.testUtils.ApiTestUtils.mockAuthenticatedRequest;
-import static org.opentripplanner.middleware.testUtils.CommonTestUtils.isEndToEnd;
+import static org.opentripplanner.middleware.testutils.ApiTestUtils.TEMP_AUTH0_USER_PASSWORD;
+import static org.opentripplanner.middleware.testutils.ApiTestUtils.mockAuthenticatedGet;
+import static org.opentripplanner.middleware.testutils.ApiTestUtils.mockAuthenticatedRequest;
+import static org.opentripplanner.middleware.testutils.CommonTestUtils.IS_END_TO_END;
 import static org.opentripplanner.middleware.utils.HttpUtils.REQUEST_METHOD.GET;
 import static org.opentripplanner.middleware.utils.HttpUtils.REQUEST_METHOD.POST;
 
@@ -51,7 +49,7 @@ public class Auth0ConnectionTest {
 
     @BeforeAll
     public static void setUp() throws IOException, InterruptedException {
-        assumeTrue(isEndToEnd);
+        assumeTrue(IS_END_TO_END);
         OtpMiddlewareTest.setUp();
         setAuthDisabled(false);
 
@@ -66,7 +64,7 @@ public class Auth0ConnectionTest {
     
     @AfterAll
     public static void tearDown() throws Auth0Exception {
-        assumeTrue(isEndToEnd);
+        assumeTrue(IS_END_TO_END);
         restoreDefaultAuthDisabled();
 
         // Delete the auth0 user created above.
