+++ src/test/java/org/opentripplanner/middleware/testutils/CommonTestUtils.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.middleware.testUtils;
+package org.opentripplanner.middleware.testutils;
 
 import org.opentripplanner.middleware.utils.FileUtils;
 
@@ -8,7 +8,7 @@ public class CommonTestUtils {
     /**
      * Whether the end-to-end environment variable is enabled.
      */
-    public static final boolean isEndToEnd = getBooleanEnvVar("RUN_E2E");
+    public static final boolean IS_END_TO_END = getBooleanEnvVar("RUN_E2E");
 
     private static final String TEST_RESOURCE_PATH = "src/test/resources/org/opentripplanner/middleware/";
 
@@ -20,11 +20,11 @@ public class CommonTestUtils {
         return variable != null && variable.equals("true");
     }
 
-    public static <T> T getResourceFileContentsAsJSON(String resourcePathName, Class<T> clazz) throws IOException {
+    public static <T> T getTestResourceAsJSON(String resourcePathName, Class<T> clazz) throws IOException {
         return FileUtils.getFileContentsAsJSON(TEST_RESOURCE_PATH + resourcePathName, clazz);
     }
 
-    public static String getResourceFileContentsAsString(String resourcePathName) throws IOException {
+    public static String getTestResourceAsString(String resourcePathName) throws IOException {
         return FileUtils.getFileContents(TEST_RESOURCE_PATH + resourcePathName);
     }
